Add save button to export node
